backend/requirements.txt
flask
flask-cors
pymongo
flask-jwt-extended
werkzeug
python-dotenv